requirements.txt
Flask==3.0.0
gunicorn==21.2.0
selenium==4.16.0
webdriver-manager==4.0.1
requests==2.32.3
beautifulsoup4==4.12.3
Pillow==11.0.0